RAG/requirements.txt
langchain_openai #For using OpenAI API
langchain_core #For embedding chatprompt template
python-dotenv
streamlit #For UI
langchain_community #For calling open source llms
langserve #For deployments, creating api routes
langchain_groq
chainlit
langchain_memory4
google-api-python-client
transformers
flask 
langchain 
langchain_groq 
python-dotenv
